requirements.txt
Django==3.0
djangorestframework==3.11.0
django-cors-headers==3.2.1
gensim==3.2.0
Janome==0.3.10
matplotlib==3.1.2
numpy==1.16.0
pandas==0.25.3
Pillow==7.0.0
psycopg2-binary==2.8.4
scikit-learn==0.22.1
urllib3==1.25.7
wordcloud==1.6.0
keras==2.2.5
tensorflow==1.15
keras-bert==0.81.0
sentencepiece==0.1.85
mecab-python3==0.996.3

django-heroku==0.3.1
gunicorn==19.9.0